Address S&K comments
