+++ client/src/components/Visualization/charts/ScatterPlot.js
@@ -9,6 +9,7 @@
 
 import { database } from '../../../firebase/firebase';
 import React, { useState, useEffect } from 'react';
+import { stringToHexColor } from '../../../utils/Utils';
 import { VictoryTheme, 
          VictoryScatter, 
          VictoryChart, 
@@ -16,32 +17,11 @@ import { VictoryTheme,
          VictoryTooltip, 
          VictoryLabel } from 'victory';
 
-/**
- * Creates a hexidecimal color string from a string input. 
- * Taken from here: `https://stackoverflow.com/a/16348977`.
- * Loosely cleaned up for style and comments.
- * This *should* return a valid hash, but if it doesn't,
- * the browser will just display `#000`.
- * @param {string} str input string 
- * @return {string}
- */
-function stringToColor(str) {
-  let hash = 0;
-  let hexColor = '#';
-
-  // Compute a hash from the input string.
-  for (let index = 0; index < str.length; index++) {
-    hash = str.charCodeAt(index) + ((hash << 5) - hash);
-  }
-  
-  // Compute a hex string from the hash.
-  for (let index = 0; index < 3; index++) {
-    let value = (hash >> (index * 8)) & 0xFF;
-    hexColor += ('00' + value.toString(16)).substr(-2);
-  }
-
-  return hexColor;
-}
+// Chart constants. 
+const chartTitleX = 175; 
+const chartTitleY = 30;
+const domainStart = 2017;
+const domainEnd = 2021;
 
 /**
  * Turn a firestore snap into a victory formatted
@@ -55,10 +35,12 @@ function formatAdvertiserCountSnapshot(snap, year) {
   const advertiser = snap.id;
 
   const victoryFormattedAdvertiserCount = 
-      {x: parseInt(year), 
-       y: parseInt(numberOfAds), 
-       fill: stringToColor(advertiser),
-       label: `${advertiser} had ${numberOfAds} ads in ${year}.`};
+      {
+        x: parseInt(year), 
+        y: parseInt(numberOfAds), 
+        fill: stringToHexColor(advertiser),
+        label: `${advertiser} had ${numberOfAds} ads in ${year}.`
+      };
 
   return victoryFormattedAdvertiserCount;
 }
@@ -79,7 +61,7 @@ function getChartRange(victoryJsonList) {
 
 /**
  * Custom react webhook. Currently calls `dev_aggregates` collection
- * and queries for the top `year` 
+ * and queries for the top `year`.
  * @param {string} year year to query
  * @param {int} limit maximum number of database documents to return
  * @return {object}
@@ -88,27 +70,27 @@ function useAdvertisers(year, queryLimit) {
   const [advertisers, setAdvertisers] = useState([]);
 
   useEffect(() => {
-    database.collection("dev_aggregates")
-            .doc(year)
-            .collection("advertisers")
-            .orderBy("numberOfAds", "desc")
-            .limit(queryLimit)
-            .get()
-            .then((snapshots) => {
-      const newAdvertisers = snapshots.docs.map(snap => {
-        return formatAdvertiserCountSnapshot(snap, year);
+      database.collection("dev_aggregates")
+              .doc(year)
+              .collection("advertisers")
+              .orderBy("numberOfAds", "desc")
+              .limit(queryLimit)
+              .get()
+              .then((snapshots) => {
+        const newAdvertisers = snapshots.docs.map(snap => {
+          return formatAdvertiserCountSnapshot(snap, year);
+        });
+
+        setAdvertisers(newAdvertisers);
       });
+    }, [advertisers, year, queryLimit]);
 
-      setAdvertisers(newAdvertisers);
-    });
-  }, []);
-
-  return advertisers;
-}
+    return advertisers;
+  }
 
 const ScatterPlot = () => {
   const queryLimit = 4; 
-
+  
   const advertisers2018 = useAdvertisers("2018", queryLimit);
   const advertisers2019 = useAdvertisers("2019", queryLimit);
   const advertisers2020 = useAdvertisers("2020", queryLimit);
@@ -123,30 +105,28 @@ const ScatterPlot = () => {
   return (
     <VictoryChart
       theme={ VictoryTheme.material }
-      domain={{ x: [2017, 2021], y: [range.min*.5, range.max*2] }}
+      domain={{ x: [domainStart, domainEnd], y: [range.min*.5, range.max*2] }}
       scale={{ y: "log" }}
     >
       <VictoryLabel
         text={ chartTitle }
-        x={ 175 } y = { 30 }
+        x={chartTitleX} y = {chartTitleY}
         textAnchor="middle"
       />
 
-      <VictoryAxis 
-        tickValues = { [2018, 2019, 2020] }
-      />
+      <VictoryAxis tickValues = {[2018, 2019, 2020]} />
 
       <VictoryAxis dependentAxis
-        tickValues = { [range.min, range.max] }
+        tickValues = {[range.min, range.max]}
         label="# of ads (log scale)"
       />
 
       <VictoryScatter
         labelComponent={ <VictoryTooltip/> }
-        labels={({ datum }) => datum.y}
+        labels={({ datum }) => datum.y }
         style={{ data: { fill: ({ datum }) => datum.fill }}}
-        size={ 10 }
-        data={ advertisers }
+        size={10}
+        data={advertisers}
       />
 
     </VictoryChart>
